add the ability to check if user is owner/admin of resource

+++ image.py
@@ -1,16 +1,15 @@
-import os
 import contextlib
+import os
 
-from flask import Blueprint, redirect, render_template, url_for
-from helpers import env
-
-from models import Image, Album
-
-from decorators import can_view
-
-from tag import get_tags, remove
+from flask import Blueprint, abort, redirect, render_template, url_for
+from flask_login import fresh_login_required
 
 from app import db
+from decorators import can_view
+from helpers import env
+from login import is_owner
+from models import Album, Image
+from tag import get_tags, remove
 
 blueprint = Blueprint('image', __name__, template_folder='templates/image')
 
@@ -24,17 +23,20 @@ def view(image_hash):
     return render_template("detail_view.html", image=image, albums=all_albums, tags=tags)
 
 @blueprint.route("<image_hash>/delete", methods=["POST"])
-@can_view
+@fresh_login_required
 def delete(image_hash):
     image = Image.query.filter_by(hash=image_hash).first()
     # Remove files
-    filename = image.hash + "." + image.extension
-    with contextlib.suppress(FileNotFoundError):
-        os.remove(os.path.join(blueprint.root_path, env("IMAGE_DIR"), filename))
-        os.remove(os.path.join(blueprint.root_path, env("THUMB_DIR"), filename))
-    # Remove from tags
-    for tag in image.tags:
-        remove(tag.namespace, tag.name, image.id)
-    db.session.delete(image)
-    db.session.commit()
-    return(redirect(url_for('home')))
+    if is_owner(image):    
+        filename = image.hash + "." + image.extension
+        with contextlib.suppress(FileNotFoundError):
+            os.remove(os.path.join(blueprint.root_path, env("IMAGE_DIR"), filename))
+            os.remove(os.path.join(blueprint.root_path, env("THUMB_DIR"), filename))
+        # Remove from tags
+        for tag in image.tags:
+            remove(tag.namespace, tag.name, image.id)
+        db.session.delete(image)
+        db.session.commit()
+        return(redirect(url_for('home')))
+    else:
+        return abort(403)
